requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.7.0
django-filter==24.3
python-decouple==3.8
psycopg2-binary==2.9.10
redis==6.4.0
celery==5.5.3
flower==2.0.1
pillow==11.3.0
channels==4.3.1
channels-redis==4.3.0
